INZ-122_resource-delete

+++ src/Database/ResourceDao.cs
@@ -48,14 +48,19 @@ internal class ResourceDao : IResourceDao
     {
         _context.Resources.Remove(resource);
     }
+    
+    
 
     public void SaveResourceGroup(DbResourceGroup group)
     {
         _context.ResourceGroups.Add(group);
     }
 
-    public void DeleteGroup(DbResourceGroup group)
+    public void DeleteGroupByGuid(Guid groupId)
     {
+        var group = _context.ResourceGroups.Query().SingleOrDefault(r => r.Guid == groupId);
+        if (group == null)
+            return;
         _context.ResourceGroups.Remove(group);
     }
 
@@ -70,6 +75,15 @@ internal class ResourceDao : IResourceDao
         return group;
     }
     
+    public DbResourceGroup GetResourceGroupById(Guid resourceGroupId)
+    {
+        var group = _context.ResourceGroups.Query()
+            .FirstOrDefault(r => r.Guid == resourceGroupId);
+        if (group == null)
+            throw new ApplicationException("Resource group not found");
+        return group;
+    }    
+        
     public IList<DbResource> GetStudentResources(int studentId)
     {
         return _context.Resources.Query()
